adding the ranking, candidate and Database

+++ backend/schema/job_description.py
@@ -9,4 +9,4 @@ class JobDescriptionData(BaseModel):
     required_skills: List[str] = Field(default=[], description="Must-have technical skills")
     preferred_skills: List[str] = Field(default=[], description="Nice-to-have skills")
     responsibilities: List[str] = Field(default=[], description="Key duties")
-    minimum_years_experience: int = Field(default=0, description="Minimum experience requirement")
+    minimum_years_experience: float = Field(default=0, description="Minimum experience requirement")
